requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
appdirs==1.4.4
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.2
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.0
discord==2.1.0
discord.py==2.1.0
exceptiongroup==1.0.4
Faker==15.3.4
frozendict==2.3.4
frozenlist==1.3.3
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.76.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
googlesearch-python==1.1.0
h11==0.14.0
html5lib==1.1
httplib2==0.21.0
idna==2.10
joblib==1.2.0
kaleido==0.2.1
lxml==4.9.2
multidict==6.0.3
multitasking==0.0.11
nltk==3.7
numpy==1.24.0
outcome==1.2.0
pandas==1.5.2
pandas-datareader==0.10.0
plotly==5.11.0
protobuf==4.21.12
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
regex==2022.10.31
requests==2.28.2
rsa==4.9
scipy==1.10.1
sec-cik-mapper==2.1.0
sec-edgar-downloader==4.3.0
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tenacity==8.1.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
webencodings==0.5.1
wsproto==1.2.0
yarl==1.8.2
yfinance==0.2.10
